DL benchmark with Callbacks - Fix Suggestion
